auth and error handling for category and product

+++ small-inventory/src/main/java/com/zihan/small_inventory/inventory/services/ProductService.java
@@ -1,11 +1,13 @@
 package com.zihan.small_inventory.inventory.services;
 
+import com.zihan.small_inventory.constants.ResponseCode;
 import com.zihan.small_inventory.inventory.items.Category;
 import com.zihan.small_inventory.inventory.items.Product;
 import com.zihan.small_inventory.inventory.repositories.CategoryRepository;
 import com.zihan.small_inventory.inventory.repositories.ProductRepository;
 import com.zihan.small_inventory.inventory.repositories.ShopRepository;
 import com.zihan.small_inventory.utils.ResponseUtil;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -25,26 +27,27 @@ public class ProductService {
     }
 
     // Create a product
-    public ResponseUtil<Product> createProduct(Product product) {
+    @PreAuthorize("#shopId == authentication.principal.shopId")
+    public ResponseUtil<Product> createProduct(String shopId, Product product) {
 
         // Check shop exists
         if (shopRepository.findByShopId(product.getShopId()).isEmpty()) {
-            return new ResponseUtil<>("Invalid shopId: shop does not exist.", 403);
+            return new ResponseUtil<>("Invalid shopId: shop does not exist.", ResponseCode.SHOP_ID_NOT_FOUND);
         }
 
         // Check category exists and belongs to this shop
         Optional<Category> categoryOpt = categoryRepository.findById(product.getCategoryId());
         if (categoryOpt.isEmpty()) {
-            return new ResponseUtil<>("Invalid categoryId: category does not exist.", 403);
+            return new ResponseUtil<>("Invalid categoryId: category does not exist.", ResponseCode.CATEGORY_NOT_FOUND);
         }
 
         if (!categoryOpt.get().getShopId().equals(product.getShopId())) {
-            return new ResponseUtil<>("Invalid categoryId: category does not exist in the given shop.", 403);
+            return new ResponseUtil<>("Invalid categoryId: category does not exist in the given shop.", ResponseCode.CATEGORY_NOT_FOUND);
         }
         // Validate fields
         String validationError = validateProduct(product);
         if (validationError != null) {
-            return new ResponseUtil<>(validationError, 403);
+            return new ResponseUtil<>(validationError, ResponseCode.PRODUCT_INVALID_INFO);
         }
         product = Product.newProduct(
                 product.getShopId(),
@@ -60,38 +63,53 @@ public class ProductService {
     }
 
     // Get all products for a shop and category
+    @PreAuthorize("#shopId == authentication.principal.shopId")
     public ResponseUtil<List<Product>> getProductsByShopAndCategory(String shopId, String categoryId) {
+        // Check category exists and belongs to this shop
+        Optional<Category> categoryOpt = categoryRepository.findById(categoryId);
+        if (categoryOpt.isEmpty()) {
+            return new ResponseUtil<>("Invalid categoryId: category does not exist.", ResponseCode.CATEGORY_NOT_FOUND);
+        }
+        if (!categoryOpt.get().getShopId().equals(categoryId)) {
+            return new ResponseUtil<>("Invalid categoryId: category does not exist in the given shop.", ResponseCode.CATEGORY_NOT_FOUND);
+        }
         return new ResponseUtil<>(productRepository.findByShopIdAndCategoryId(shopId, categoryId));
     }
 
     // Get product by ID
-    public ResponseUtil<Product> getProductById(String productId) {
+    @PreAuthorize("#shopId == authentication.principal.shopId")
+    public ResponseUtil<Product> getProductById(String shopId, String productId) {
         Optional<Product> product = productRepository.findById(productId);
         if (product.isEmpty()) {
-            return new ResponseUtil<>("Product not found.", 403);
+            return new ResponseUtil<>("Product not found.", ResponseCode.PRODUCT_NOT_FOUND);
+        }
+        if (!product.get().getShopId().equals(shopId)){
+            return new ResponseUtil<>("Product not found for this shop.", ResponseCode.PRODUCT_NOT_FOUND);
         }
         return new ResponseUtil<>(product.get());
     }
     // Update product
-    public ResponseUtil<Product> updateProduct(Product product) {
+    @PreAuthorize("#shopId == authentication.principal.shopId")
+    public ResponseUtil<Product> updateProduct( String shopId, Product product) {
         Optional<Product> existing = productRepository.findById(product.getProductId());
         if (existing.isEmpty()) {
-            return new ResponseUtil<>("Product not found", 403);
+            return new ResponseUtil<>("Product not found", ResponseCode.PRODUCT_NOT_FOUND);
+        }
+        if (!existing.get().getShopId().equals(shopId)){
+            return new ResponseUtil<>("Product not found for this shop.", ResponseCode.PRODUCT_NOT_FOUND);
         }
         // Check category exists and belongs to this shop
         Optional<Category> categoryOpt = categoryRepository.findById(product.getCategoryId());
         if (categoryOpt.isEmpty()) {
-            return new ResponseUtil<>("Invalid categoryId: category does not exist.", 403);
+            return new ResponseUtil<>("Invalid categoryId: category does not exist.", ResponseCode.CATEGORY_NOT_FOUND);
         }
-        System.out.println(categoryOpt.get().getShopId());
-        System.out.println(product.getShopId());
         if (!categoryOpt.get().getShopId().equals(product.getShopId())) {
-            return new ResponseUtil<>("Invalid categoryId: category does not exist in the given shop.", 403);
+            return new ResponseUtil<>("Invalid categoryId: category does not exist in the given shop.", ResponseCode.CATEGORY_NOT_FOUND);
         }
 
         String validationError = validateProduct(product);
         if (validationError != null) {
-            return new ResponseUtil<>(validationError, 403);
+            return new ResponseUtil<>(validationError, ResponseCode.PRODUCT_INVALID_INFO);
         }
 
         Product updated = existing.get();
@@ -108,14 +126,17 @@ public class ProductService {
     }
 
     // Delete product
-    public ResponseUtil<String> deleteProduct(String productId) {
+    @PreAuthorize("#shopId == authentication.principal.shopId")
+    public ResponseUtil<String> deleteProduct(String shopId, String productId) {
         Optional<Product> existing = productRepository.findById(productId);
         if (existing.isEmpty()) {
-            return new ResponseUtil<>("Product not found", 403);
+            return new ResponseUtil<>("Product not found", ResponseCode.PRODUCT_NOT_FOUND);
         }
-
+        if (!existing.get().getShopId().equals(shopId)){
+            return new ResponseUtil<>("Product not found for this shop.", ResponseCode.PRODUCT_NOT_FOUND);
+        }
         productRepository.delete(productId);
-        return new ResponseUtil<>("Product deleted successfully.", 200);
+        return new ResponseUtil<>("Product deleted successfully.", ResponseCode.SUCCESS);
     }
 
     // Helper to validate product fields
